v3.0.0: The rework
This project has now transformed into a fully-fledged website for
managing the schedule. It now runs on Rocket and has a nice FullCalendar
ui with the ability to login and create and remove custom events.

We have also updated to use JouwLoon 2.0, this makes a lot of things a
lot easier (we finally have a normal API, not requiring 100's of web
requests for a single schedule). However, programmatic login has become
a little more complicated due to the server-side switch to ASP.NET
WebForms.